Backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
python-multipart>=0.0.17
python-dotenv>=1.0.1
alembic>=1.14.0
email-validator>=2.2.0
websockets>=14.0
redis>=5.0.0
openpyxl>=3.1.0
httpx>=0.28.0
slowapi>=0.1.9
pytest>=8.0.0
